Creadas las funciones de Borrar y Modificar

+++ app/src/main/java/com/svalero/cybershopapp/domain/Client.java
@@ -5,6 +5,7 @@ import androidx.room.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import lombok.NonNull;
 
 @Data
 @AllArgsConstructor
@@ -13,7 +14,7 @@ import lombok.NoArgsConstructor;
 public class Client {
 
     @PrimaryKey(autoGenerate = true)
-    private int id;
+    private long id;
     @ColumnInfo
     private String name;
     @ColumnInfo
@@ -33,11 +34,11 @@ public class Client {
         this.vip = vip;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
